style: Applied automatic cargo clippy fixes
Just applied auto fixes from cargo clippy --fix (then a follow up auto
format using cargo fmt). Verified that all tests work.

+++ crates/core/src/prover/external/backends/vampire/integrated.rs
@@ -134,7 +134,7 @@ impl ProverRunner for IntegratedVampireRunner {
         // loop varies it run-to-run); fall back to the runner's own field.
         let timeout = opts.timeout();
         if timeout > 0 {
-            vp_opts.timeout(Duration::from_secs(timeout as u64));
+            vp_opts.timeout(Duration::from_secs(timeout));
         }
         match opts.mode {
             ProverMode::Prove => {
